Added update for question type api endpoint

+++ pdf-app/src/main/java/com/bitMiners/pdf/services/impl/QuestionTypeServiceImpl.java
@@ -1,12 +1,14 @@
 package com.bitMiners.pdf.services.impl;
 
 import com.bitMiners.pdf.domain.QuestionType;
+import com.bitMiners.pdf.exceptions.PdfApiException;
 import com.bitMiners.pdf.repositories.QuestionTypeRepository;
 import com.bitMiners.pdf.services.QuestionTypeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class QuestionTypeServiceImpl implements QuestionTypeService {
@@ -19,10 +21,18 @@ public class QuestionTypeServiceImpl implements QuestionTypeService {
     }
 
     public QuestionType updateQuestionType(QuestionType questionType) {
+        QuestionType dbRecord = this.questionTypeRepository.findOne(questionType.getId());
+        if (Objects.isNull(dbRecord)) {
+            throw new PdfApiException("No Question Type found with id: " + questionType.getId(), 404);
+        }
         return questionTypeRepository.update(questionType);
     }
 
     public void deleteQuestionType(Long id) {
+        QuestionType dbRecord = this.questionTypeRepository.findOne(id);
+        if (Objects.isNull(dbRecord)) {
+            throw new PdfApiException("No Question Type found with id: " + id, 404);
+        }
         questionTypeRepository.delete(id);
     }
